Move configurator fixtures to conftest
