--- tests/unit/test_configurator/fixtures/extension.py
@@ -1,17 +1,0 @@
-# © 2023 SolarWinds Worldwide, LLC. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at:http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
-import pytest
-
-@pytest.fixture(name="mock_extension")
-def mock_extension(mocker):
-    return mocker.patch(
-        "solarwinds_apm.extension"
-    )
-
-@pytest.fixture(name="mock_reporter")
-def mock_reporter(mocker):
-    return mocker.Mock()
