+++ tests/unit/test_configurator/test_configurator_traces_exporter.py
@@ -10,19 +10,7 @@ import pytest
 from solarwinds_apm import configurator
 
 # otel fixtures
-from .fixtures.batch_span_processor import mock_bsprocessor
 from .fixtures.trace import get_trace_mocks
-
-# apm python fixtures
-from .fixtures.apm_config import (
-    mock_apmconfig_disabled,
-    mock_apmconfig_enabled,
-)
-from .fixtures.extension import (
-    mock_reporter,
-)
-from .fixtures.fwkv_manager import mock_fwkv_manager
-from .fixtures.txn_name_manager import mock_txn_name_manager
 
 
 class TestConfiguratorTracesExporter:
